Feat: create toolbar and searchbar
