+++ src/components/GlobalNavigationBar/GlobalNavigationBar.module.scss
@@ -1,4 +1,4 @@
-@use "@/sass";
+@use "@/sass" as *;
 
 .container {
   position: fixed;
@@ -13,6 +13,9 @@
   box-shadow:
     0px -2px 8px 0px rgba(20, 20, 20, 0.04),
     0px -1px 1px 0px rgba(20, 20, 20, 0.02);
+
+  background-color: $neutral0;
+
   a {
     flex-grow: 1;
