--- src/components/Home/RecomendedItems/RecommendedLocation/RecommendedLocation.module.scss
@@ -1,18 +1,0 @@
-@use "@/sass";
-
-.container {
-  min-width: 122px;
-  height: 122px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  border-radius: 122px;
-
-  color: sass.$neutral0;
-
-  background: sass.$primary200;
-
-  @include sass.typography(titleMedium);
-}
