+++ src/components/Home/RecommendedItemList/RecommendedItemList.module.scss
@@ -2,7 +2,7 @@
 
 .container {
   display: flex;
-  gap: 8px;
+  gap: 16px;
 
   padding: 0 20px;
