+++ src/components/Home/RecommendedLocationList/RecommendedLocationList.module.scss
