+++ src/components/Home/TabBar/TabBar.module.scss
@@ -1,0 +1,22 @@
+@use "@/sass" as *;
+
+.container {
+  position: fixed;
+  top: 0;
+
+  width: 100%;
+  max-width: 50rem;
+  height: 5.6rem;
+
+  padding: 20px 16px;
+
+  background-color: $neutral0;
+
+  .icons {
+    display: flex;
+    justify-content: flex-end;
+    gap: 12px;
+
+    font-size: 2.4rem;
+  }
+}
