+++ src/components/Home/VoteAtHome/VoteCard/CardNull/CardNull.module.scss
@@ -1,8 +1,8 @@
-@use "@/sass";
+@use "@/sass" as *;
 
 .vote_box {
   width: 100%;
-  height: 175px;
+  height: 17.5rem;
 
   padding: 0 20px;
 
@@ -21,7 +21,7 @@
       rgba(98, 170, 255, 0.3) 142.38%
     );
 
-    border-radius: 16px;
+    border-radius: 1.6rem;
 
     .left {
       height: 100%;
@@ -31,39 +31,12 @@
       justify-content: space-between;
 
       .text {
-        @include sass.typography(titleSmall);
+        @include typography(titleSmall);
       }
     }
     .right {
-      width: 98px;
-      height: 98px;
-      padding: 5px 13px;
-    }
-  }
-  .contents {
-    height: 100%;
-
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    padding: 33px 24px;
-
-    .left {
-      height: 100%;
-
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-
-      .text {
-        @include sass.typography(titleSmall);
-        color: sass.$neutral0;
-      }
-    }
-    .right {
-      width: 98px;
-      height: 98px;
+      width: 9.8rem;
+      height: 9.8rem;
       padding: 5px 13px;
     }
   }
