+++ src/pages/Home/Home.module.scss
@@ -1,13 +1,15 @@
-@use "@/sass";
+@use "@/sass" as *;
 
 .container {
   width: 100%;
 
   display: flex;
   flex-direction: column;
-  gap: 32px;
+  gap: 3.2rem;
 
-  padding-bottom: 56px;
+  // gnb + 홈 하단여백
+  padding-top: 56px;
+  padding-bottom: 75px;
 
   -ms-overflow-style: none;
   scrollbar-width: none;
@@ -16,35 +18,27 @@
   }
 
   .title {
-    @include sass.typography(titleMedium);
+    @include typography(titleMedium);
     color: #1d2433;
 
-    padding: 8px 0;
-  }
-
-  .section {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
+    padding: 8px 20px;
   }
 
   .lists_box {
     display: flex;
     flex-direction: column;
-    gap: 48px;
+    gap: 4.8rem;
   }
 
-  .item_list_8px {
+  .column_4px {
     display: flex;
-    gap: 8px;
-
-    overflow-x: scroll;
+    flex-direction: column;
+    gap: 0.4rem;
   }
 
-  .item_list_16px {
+  .column_8px {
     display: flex;
-    gap: 16px;
-
-    overflow-x: scroll;
+    flex-direction: column;
+    gap: 0.8rem;
   }
 }
